GEODE-3436: Restore refactoring of ConfigCommands
* See initial commit GEODE-3254 (97c4e9a59f17c7bc914e39dd048b0a4cd96293c4)
